[Modify] Game is working
Use pair not Point struct.

In MakeBoard : add conditions to make board length even, and < 20

In InputXY : get Input from user and check it's condition valid

In CheckCondition : check 3 condition that we write on Init.md
In here, need to clear the vector valid    to put stone.

In ShowValid : you have to add status == "3"  for the vacant.
need to use SetBoard to chage the original Board.

In ToggleStone : need to Toggle the (x, y) stone.